openai_v2: move responses wrappers into patch.py
